Display more digits on LoopSlider's labels
