+++ lib/widgets.dart
@@ -81,10 +81,3 @@ class LoadingScreen extends StatelessWidget {
     );
   }
 }
-
-String durationString(Duration duration) {
-  return RegExp(r'((^0*[1-9]\d*:)?\d{2}:\d{2})\.\d+$')
-          .firstMatch("$duration")
-          ?.group(1) ??
-      "$duration";
-}
